requirements.txt
paho-mqtt==2.1.0
pi1wire==0.3.0
